Squirrels
The squirrels are lying low. Maybe it's the early morning surprise
attack I launched – my four year old let out of the backdoor as they yet
again ate the bread for the birds. Maybe they're regrouping in the
spotted laurel, planning their next sortie. Either way, I know, with the
weariness of the besieged, that it won't be long before battle is
rejoined.
"They're very smart, they work things out, and chew their way into
people's houses and suss out how to get exactly what they want," says
Ian Woods of Grey Squirrel Control. There are estimated to be more than
2.5 million squirrels in England, Scotland and Wales. Last summer, with
the warm weather encouraging us to leave our kitchen door open, the
low-level dirty war that the squirrels had been waging on our house
(gnawing a hole through the shed and ripping open the binbags
buffet-style) stepped up a notch. I began to find them in the kitchen,
rummaging through the bread bin.
Then they discovered the bathroom skylight and started coming in through
that. Battle culminated with me and one particularly stubborn little
fighter facing off on the stairs, trying to work out who was going to
get out of the way first (me, of course).
So what can you do? Woods says that more and more people are buying
traps from him. "However, it's illegal to re-release squirrels into the
wild. So if you use a live trap, you need to kill the squirrel yourself
and the only legal ways of doing that are shooting it with an air rifle,
or putting it in a sack and hitting it on the head."
In other words, they've got us, chaps. My back door will be staying shut
while I contemplate the terms of surrender.